+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 2026-01-29
+
+**Changed**
+* Add results of OUTRIDER for all candidate exons.
+* Add results of FRASER genome-wide (as a Manhattan plot & data.table
+
+
 ### 2026-01-28
 
 **Changed**